feat: Display resource hub item's icon and subtitle on space page
