+++ assets/js/features/ResourceHub/index.tsx
@@ -1,3 +1,4 @@
 export { NodesList } from "./NodesList";
 export { ZeroNodes } from "./ZeroNodes";
 export { AddFilesButtonAndForms } from "./AddNewFiles";
+export { findIcon, findSubtitle, NodeType } from "./utils";
