+++ lib/operately_web/api/queries/list_space_tools.ex
@@ -5,7 +5,7 @@ defmodule OperatelyWeb.Api.Queries.ListSpaceTools do
   alias Operately.Goals.Goal
   alias Operately.Projects.Project
   alias Operately.Messages.{MessagesBoard, Message}
-  alias Operately.ResourceHubs.ResourceHub
+  alias Operately.ResourceHubs.{ResourceHub, Node}
   alias Operately.Groups.SpaceTools
   alias Operately.Access.Filters
 
@@ -88,13 +88,20 @@ defmodule OperatelyWeb.Api.Queries.ListSpaceTools do
   end
 
   defp load_resource_hubs(space_id, me) do
+    nodes_q =
+      from(n in Node,
+        where: is_nil(n.parent_folder_id),
+        preload: [document: :node, folder: :node, file: [:node, :preview_blob, :blob]]
+      )
+
     hubs =
       from(h in ResourceHub,
-        preload: :nodes,
+        preload: [nodes: ^nodes_q],
         where: h.space_id == ^space_id
       )
       |> Filters.filter_by_view_access(me.id)
       |> Repo.all()
+      |> ResourceHub.set_children_count()
 
     {:ok, hubs}
   end
